Setting up battery-widget
Using the battery-widget from [awesome-wm-widgets][1], which is now a
submodule of this repo. Within that submodule, I only do a sparse
checkout of the widgets that I want to use (so far, battery-widget and
volume-widget).

Also changed the warning message (`warning_msg_title`) used by the
battery-widget (partially to practise using a Lua table).

battery-widget depends on `[arc-icon-theme-git][2]` and
`[moka-icon-theme-git][3]`, both of which were installed from the AUR.

 [1]: https://github.com/streetturtle/awesome-wm-widgets
 [2]: https://aur.archlinux.org/packages/arc-icon-theme-git/
 [3]: https://aur.archlinux.org/packages/moka-icon-theme-git/